backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
playwright==1.42.0
beautifulsoup4==4.12.3
html2text==2024.2.26
google-genai
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
playwright-stealth==1.0.6
